create Sidebar component
